Add `-w` to buildx ldflags This disables DWARF generation, which dramatically decreases the size of the binary.
Notably, this does *not* include `-s` which disables the symbol table, as doing so would *also* make it so that `govulncheck` cannot be run on the resulting binary with meaningful results.

Before is ~90M and after is ~72M.

If we add `-s` as well, that ~72M drops to ~63M, but we also lose `govulncheck` so again, IMO, that's not worth doing.

Signed-off-by: Teddysun <i@teddysun.com>